requirements.txt
numpy
pandas
fastapi
uvicorn
tensorflow
python-dotenv
python-multipart
pillow
